+++ OneBlink.SDK.Tests/EmailTemplatesClientTests.cs
@@ -13,6 +13,7 @@ namespace OneBlink.SDK.Tests
         private string SECRET_KEY;
         private long formsAppEnvironmentId = 22;
         private string organisationId = "5c58beb2ff59481100000002";
+        private long workspaceId = 12;
         public EmailTemplatesClientTests()
         {
             bool ignoreExceptions = true;
@@ -28,6 +29,11 @@ namespace OneBlink.SDK.Tests
             if (!String.IsNullOrWhiteSpace(organisationId))
             {
                 this.organisationId = organisationId;
+            }
+            string workspaceId = Environment.GetEnvironmentVariable("WORKSPACE_ID");
+            if (!String.IsNullOrWhiteSpace(workspaceId))
+            {
+                this.workspaceId = long.Parse(workspaceId);
             }
         }
         [Fact]
@@ -59,12 +65,12 @@ namespace OneBlink.SDK.Tests
             newEmailTemplate.name = "Unit test environment";
             newEmailTemplate.type = "FORM_SUBMISSION_EVENT_PDF";
             newEmailTemplate.organisationId = organisationId;
+            newEmailTemplate.workspaceId = workspaceId;
             newEmailTemplate.environments = new System.Collections.Generic.List<EmailTemplateEnvironment>()
             {
                 new EmailTemplateEnvironment()
                 {
                     formsAppEnvironmentId = formsAppEnvironmentId,
-                    template = "Created via unit test"
                 }
             };