+++ OneBlink.SDK/models/EmailTemplate.cs
@@ -34,6 +34,10 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        public long workspaceId
+        {
+            get; set;
+        }
 
         public string type
         {