+++ OneBlink.SDK/models/TaskGroup.cs
@@ -34,5 +34,9 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        public long workspaceId
+        {
+            get; set;
+        }
     }
 }