+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- `workspaceId` to `Form`
+- `workspaceId` to `EmailTemplate`, `Form`, `FormElementList`, `FormElementLookup`, `FormsAppBase`, `Task` and `TaskGroup`
 - Optional parameter `workspaceId` to `EmailTemplatesClient.Search()`, `FormElementListsClient.Search()`, `FormElementLookupsClient.Search()`, `FormsAppEnvironmentsClient.Search()` and `FormsClient.Search()`
 
 ## [11.2.1] - 2026-02-23